Split `TopLevelInformationController`
We split `TopLevelInformationController` into `TopLevelInformationController` and `ScopeController`, with the former concerned with the name of the profession, and the later concerned with only the nations and the sectors
