+++ cypress/integration/admin/professions/edit.spec.ts
@@ -60,6 +60,19 @@ describe('Editing an existing profession', () => {
         'Registered Trademark Attorney',
       );
 
+      cy.translate('nations.england').then((england) => {
+        cy.translate('nations.scotland').then((scotland) => {
+          cy.translate('nations.wales').then((wales) => {
+            cy.translate('nations.northernIreland').then((northernIreland) => {
+              cy.checkSummaryListRowList(
+                'professions.form.label.scope.nations',
+                [england, scotland, wales, northernIreland],
+              );
+            });
+          });
+        });
+      });
+
       cy.checkSummaryListRowValue(
         'professions.form.label.regulatoryBody.regulatedAuthority',
         'Law Society of England and Wales',
@@ -91,7 +104,7 @@ describe('Editing an existing profession', () => {
 
       // Conditional radio buttons add an additional `aria-expanded` field,
       // so ignore that rule on this page
-      cy.checkAccessibility({ 'aria-allowed-attr': { enabled: false } });
+      cy.checkAccessibility();
 
       cy.translate('professions.form.captions.edit').then((editCaption) => {
         cy.get('body').contains(editCaption);
@@ -104,6 +117,39 @@ describe('Editing an existing profession', () => {
         'professions.form.label.topLevelInformation.name',
         'Updated name',
       );
+
+      cy.clickSummaryListRowAction(
+        'professions.form.label.scope.nations',
+        'Change',
+      );
+
+      // Conditional radio buttons add an additional `aria-expanded` field,
+      // so ignore that rule on this page
+      cy.checkAccessibility({ 'aria-allowed-attr': { enabled: false } });
+
+      cy.translate('professions.form.label.scope.certainNations').then(
+        (certainNations) => {
+          cy.get('label').contains(certainNations).click();
+        },
+      );
+      cy.translate('nations.england').then((england) => {
+        cy.get('label').contains(england).click();
+      });
+      cy.translate('app.continue').then((buttonText) => {
+        cy.get('button').contains(buttonText).click();
+      });
+
+      cy.translate('nations.scotland').then((scotland) => {
+        cy.translate('nations.wales').then((wales) => {
+          cy.translate('nations.northernIreland').then((northernIreland) => {
+            cy.checkSummaryListRowList('professions.form.label.scope.nations', [
+              scotland,
+              wales,
+              northernIreland,
+            ]);
+          });
+        });
+      });
 
       cy.clickSummaryListRowAction(
         'professions.form.label.regulatoryBody.regulatedAuthority',
@@ -264,7 +310,7 @@ describe('Editing an existing profession', () => {
       });
 
       cy.clickSummaryListRowAction(
-        'professions.form.label.topLevelInformation.nations',
+        'professions.form.label.scope.nations',
         'Change',
       );
 
@@ -300,15 +346,15 @@ describe('Editing an existing profession', () => {
       });
 
       cy.clickSummaryListRowAction(
-        'professions.form.label.topLevelInformation.nations',
+        'professions.form.label.scope.nations',
         'Change',
       );
 
-      cy.translate(
-        'professions.form.label.topLevelInformation.certainNations',
-      ).then((certainNations) => {
-        cy.get('label').contains(certainNations).click();
-      });
+      cy.translate('professions.form.label.scope.certainNations').then(
+        (certainNations) => {
+          cy.get('label').contains(certainNations).click();
+        },
+      );
 
       cy.get('[type="checkbox"]').uncheck('GB-SCT');
 
@@ -360,7 +406,7 @@ describe('Editing an existing profession', () => {
       });
 
       cy.clickSummaryListRowAction(
-        'professions.form.label.topLevelInformation.name',
+        'professions.form.label.scope.nations',
         'Change',
       );
 
@@ -377,7 +423,7 @@ describe('Editing an existing profession', () => {
           cy.translate('nations.wales').then((wales) => {
             cy.translate('nations.northernIreland').then((northernIreland) => {
               cy.checkSummaryListRowList(
-                'professions.form.label.topLevelInformation.nations',
+                'professions.form.label.scope.nations',
                 [england, scotland, wales, northernIreland],
               );
             });
