+++ cypress/integration/admin/professions/new.spec.ts
@@ -35,10 +35,7 @@ describe('Adding a new profession', () => {
         cy.get('button').contains(buttonText).click();
       });
 
-      // Conditional radio buttons add an additional `aria-expanded` field,
-      // so ignore that rule on this page
-      cy.checkAccessibility({ 'aria-allowed-attr': { enabled: false } });
-
+      cy.checkAccessibility();
       cy.translate('professions.form.headings.topLevelInformation').then(
         (heading) => {
           cy.get('body').should('contain', heading);
@@ -48,13 +45,19 @@ describe('Adding a new profession', () => {
         cy.get('body').contains(addCaption);
       });
       cy.get('input[name="name"]').type('Example Profession');
-
-      cy.translate(
-        'professions.form.label.topLevelInformation.certainNations',
-      ).then((certainNations) => {
-        cy.get('label').contains(certainNations).click();
-        cy.get('[type="checkbox"]').check('GB-ENG');
+      cy.translate('app.continue').then((buttonText) => {
+        cy.get('button').contains(buttonText).click();
       });
+
+      // Conditional radio buttons add an additional `aria-expanded` field,
+      // so ignore that rule on this page
+      cy.checkAccessibility({ 'aria-allowed-attr': { enabled: false } });
+      cy.translate('professions.form.label.scope.certainNations').then(
+        (certainNations) => {
+          cy.get('label').contains(certainNations).click();
+          cy.get('[type="checkbox"]').check('GB-ENG');
+        },
+      );
 
       cy.translate('industries.constructionAndEngineering').then(
         (constructionAndEngineering) => {
@@ -207,14 +210,14 @@ describe('Adding a new profession', () => {
 
       cy.translate('nations.england').then((england) => {
         cy.checkSummaryListRowValue(
-          'professions.form.label.topLevelInformation.nations',
+          'professions.form.label.scope.nations',
           england,
         );
         cy.get('body').should('contain', england);
       });
 
       cy.checkSummaryListRowValue(
-        'professions.form.label.topLevelInformation.industry',
+        'professions.form.label.scope.industry',
         'Construction & Engineering',
       );
