+++ src/professions/admin/interfaces/top-level-details.template.ts
@@ -1,11 +1,5 @@
-import { CheckboxItems } from '../../../common/interfaces/checkbox-items.interface';
-import { CheckboxArgs } from '../../../common/interfaces/checkbox-args.interface';
-
 export interface TopLevelDetailsTemplate {
   name: string | null;
-  coversUK: boolean | null;
-  industriesCheckboxItems: CheckboxItems[];
-  nationsCheckboxArgs: CheckboxArgs;
   captionText: string;
   change: boolean;
   errors: object | undefined;
