+++ src/professions/admin/regulatory-body.controller.ts
@@ -42,7 +42,7 @@ export class RegulatoryBodyController {
   @BackLink((request: Request) =>
     request.query.change === 'true'
       ? '/admin/professions/:professionId/versions/:versionId/check-your-answers'
-      : '/admin/professions/:professionId/versions/:versionId/top-level-information/edit',
+      : '/admin/professions/:professionId/versions/:versionId/scope/edit',
   )
   async edit(
     @Res() res: Response,
@@ -68,7 +68,7 @@ export class RegulatoryBodyController {
   @BackLink((request: Request) =>
     request.body.change === 'true'
       ? '/admin/professions/:professionId/versions/:versionId/check-your-answers'
-      : '/admin/professions/:professionId/versions/:versionId/top-level-information/edit',
+      : '/admin/professions/:professionId/versions/:versionId/scope/edit',
   )
   async update(
     @Res() res: Response,
