+++ src/professions/admin/top-level-information.controller.spec.ts
@@ -1,43 +1,23 @@
 import { DeepMocked, createMock } from '@golevelup/ts-jest';
 import { TestingModule, Test } from '@nestjs/testing';
 import { Response } from 'express';
-import { I18nService } from 'nestjs-i18n';
 import professionFactory from '../../testutils/factories/profession';
 
-import { IndustriesService } from '../../industries/industries.service';
 import { ProfessionsService } from '../professions.service';
 import { TopLevelInformationController } from './top-level-information.controller';
-import industryFactory from '../../testutils/factories/industry';
-import { createMockI18nService } from '../../testutils/create-mock-i18n-service';
-import { translationOf } from '../../testutils/translation-of';
-import { ProfessionVersionsService } from '../profession-versions.service';
-import professionVersionFactory from '../../testutils/factories/profession-version';
 
 describe('TopLevelInformationController', () => {
   let controller: TopLevelInformationController;
   let professionsService: DeepMocked<ProfessionsService>;
-  let professionVersionsService: DeepMocked<ProfessionVersionsService>;
-  let industriesService: DeepMocked<IndustriesService>;
   let response: DeepMocked<Response>;
-  let i18nService: DeepMocked<I18nService>;
 
   beforeEach(async () => {
-    industriesService = createMock<IndustriesService>();
     professionsService = createMock<ProfessionsService>();
-    professionVersionsService = createMock<ProfessionVersionsService>();
-
-    i18nService = createMockI18nService();
 
     const module: TestingModule = await Test.createTestingModule({
       controllers: [TopLevelInformationController],
       providers: [
-        { provide: IndustriesService, useValue: industriesService },
         { provide: ProfessionsService, useValue: professionsService },
-        {
-          provide: ProfessionVersionsService,
-          useValue: professionVersionsService,
-        },
-        { provide: I18nService, useValue: i18nService },
       ],
     }).compile();
 
@@ -49,191 +29,39 @@ describe('TopLevelInformationController', () => {
   });
 
   describe('edit', () => {
-    describe('when editing a just-created, blank Profession and its blank Version', () => {
-      it('should fetch all Industries and Nations to be displayed in an option select, with none of them checked', async () => {
+    describe('when editing a just-created, blank Profession', () => {
+      it('should show a blank name', async () => {
         const blankProfession = professionFactory
           .justCreated('profession-id')
           .build();
 
-        const blankVersion = professionVersionFactory
-          .justCreated('version-id')
-          .build({ profession: blankProfession });
-
         professionsService.findWithVersions.mockResolvedValue(blankProfession);
-        professionVersionsService.findWithProfession.mockResolvedValue(
-          blankVersion,
-        );
 
-        const industries = [
-          industryFactory.build({
-            name: 'industries.health',
-            id: 'health-uuid',
-          }),
-          industryFactory.build({
-            name: 'industries.constructionAndEngineering',
-            id: 'construction-uuid',
-          }),
-        ];
-
-        industriesService.all.mockResolvedValue(industries);
-
-        await controller.edit(response, 'profession-id', 'version-id', 'false');
+        await controller.edit(response, 'profession-id', 'false');
 
         expect(response.render).toHaveBeenCalledWith(
           'admin/professions/top-level-information',
           expect.objectContaining({
             name: undefined,
-            coversUK: null,
-            industriesCheckboxItems: [
-              {
-                text: translationOf('industries.health'),
-                value: 'health-uuid',
-                checked: false,
-              },
-              {
-                text: translationOf('industries.constructionAndEngineering'),
-                value: 'construction-uuid',
-                checked: false,
-              },
-            ],
-            nationsCheckboxArgs: {
-              idPrefix: 'nations',
-              name: 'nations[]',
-              hint: {
-                text: translationOf('professions.form.checkboxes.hint'),
-              },
-              items: [
-                {
-                  text: translationOf('nations.england'),
-                  value: 'GB-ENG',
-                  checked: false,
-                },
-                {
-                  text: translationOf('nations.scotland'),
-                  value: 'GB-SCT',
-                  checked: false,
-                },
-                {
-                  text: translationOf('nations.wales'),
-                  value: 'GB-WLS',
-                  checked: false,
-                },
-                {
-                  text: translationOf('nations.northernIreland'),
-                  value: 'GB-NIR',
-                  checked: false,
-                },
-              ],
-            },
           }),
         );
-        expect(industriesService.all).toHaveBeenCalled();
       });
     });
 
-    describe('when an existing Profession and Version are found', () => {
-      it('should pre-fill the Profession name and pre-select checkboxes for selected Nations and Industries', async () => {
-        const healthIndustry = industryFactory.build({
-          name: 'industries.health',
-          id: 'health-uuid',
-        });
-
+    describe('when an existing Profession is found', () => {
+      it('should pre-fill the Profession name', async () => {
         const profession = professionFactory.build({
           name: 'Example Profession',
         });
 
-        const version = professionVersionFactory.build({
-          id: 'version-id',
-          profession: profession,
-          occupationLocations: ['GB-ENG', 'GB-SCT'],
-          industries: [healthIndustry],
-        });
-
         professionsService.findWithVersions.mockResolvedValue(profession);
-        professionVersionsService.findWithProfession.mockResolvedValue(version);
 
-        industriesService.all.mockResolvedValue([
-          healthIndustry,
-          industryFactory.build({
-            name: 'industries.constructionAndEngineering',
-            id: 'construction-uuid',
-          }),
-        ]);
-
-        await controller.edit(response, 'profession-id', 'version-id', 'false');
+        await controller.edit(response, 'profession-id', 'false');
 
         expect(response.render).toHaveBeenCalledWith(
           'admin/professions/top-level-information',
           expect.objectContaining({
             name: 'Example Profession',
-            coversUK: false,
-            industriesCheckboxItems: [
-              {
-                text: translationOf('industries.health'),
-                value: 'health-uuid',
-                checked: true,
-              },
-              {
-                text: translationOf('industries.constructionAndEngineering'),
-                value: 'construction-uuid',
-                checked: false,
-              },
-            ],
-            nationsCheckboxArgs: {
-              idPrefix: 'nations',
-              name: 'nations[]',
-              hint: {
-                text: translationOf('professions.form.checkboxes.hint'),
-              },
-              items: [
-                {
-                  text: translationOf('nations.england'),
-                  value: 'GB-ENG',
-                  checked: true,
-                },
-                {
-                  text: translationOf('nations.scotland'),
-                  value: 'GB-SCT',
-                  checked: true,
-                },
-                {
-                  text: translationOf('nations.wales'),
-                  value: 'GB-WLS',
-                  checked: false,
-                },
-                {
-                  text: translationOf('nations.northernIreland'),
-                  value: 'GB-NIR',
-                  checked: false,
-                },
-              ],
-            },
-          }),
-        );
-        expect(industriesService.all).toHaveBeenCalled();
-      });
-
-      it('should set coversUK to 1 if the profession covers all of the UK', async () => {
-        const profession = professionFactory.build({
-          name: 'Example Profession',
-        });
-
-        const version = professionVersionFactory.build({
-          id: 'version-id',
-          profession: profession,
-          occupationLocations: ['GB-ENG', 'GB-SCT', 'GB-WLS', 'GB-NIR'],
-          industries: [],
-        });
-
-        professionsService.findWithVersions.mockResolvedValue(profession);
-        professionVersionsService.findWithProfession.mockResolvedValue(version);
-
-        await controller.edit(response, 'profession-id', 'version-id', 'false');
-
-        expect(response.render).toHaveBeenCalledWith(
-          'admin/professions/top-level-information',
-          expect.objectContaining({
-            coversUK: true,
           }),
         );
       });
@@ -247,27 +75,11 @@ describe('TopLevelInformationController', () => {
           .justCreated('profession-id')
           .build();
 
-        const version = professionVersionFactory.build({
-          id: 'version-id',
-          profession: profession,
-        });
-
         professionsService.findWithVersions.mockResolvedValue(profession);
-        professionVersionsService.findWithProfession.mockResolvedValue(version);
 
         const topLevelDetailsDto = {
           name: 'Gas Safe Engineer',
-          coversUK: '0',
-          nations: ['GB-ENG'],
-          industries: ['construction-uuid'],
         };
-
-        const constructionIndustry = industryFactory.build({
-          name: 'industries.constructionAndEngineering',
-          id: 'construction-uuid',
-        });
-
-        industriesService.findByIds.mockResolvedValue([constructionIndustry]);
 
         await controller.update(
           topLevelDetailsDto,
@@ -283,50 +95,8 @@ describe('TopLevelInformationController', () => {
           }),
         );
 
-        expect(professionVersionsService.save).toHaveBeenCalledWith(
-          expect.objectContaining({
-            id: 'version-id',
-            occupationLocations: ['GB-ENG'],
-            industries: [constructionIndustry],
-            profession: profession,
-          }),
-        );
-
         expect(response.redirect).toHaveBeenCalledWith(
-          '/admin/professions/profession-id/versions/version-id/regulatory-body/edit',
-        );
-      });
-
-      it('sets all nations when `coversUK` is set to `1`', async () => {
-        const profession = professionFactory
-          .justCreated('profession-id')
-          .build();
-
-        const version = professionVersionFactory.build({
-          id: 'version-id',
-          profession: profession,
-        });
-
-        professionsService.findWithVersions.mockResolvedValue(profession);
-        professionVersionsService.findWithProfession.mockResolvedValue(version);
-
-        const topLevelDetailsDto = {
-          name: 'Gas Safe Engineer',
-          coversUK: '1',
-          industries: ['construction-uuid'],
-        };
-
-        await controller.update(
-          topLevelDetailsDto,
-          response,
-          'profession-id',
-          'version-id',
-        );
-
-        expect(professionVersionsService.save).toHaveBeenCalledWith(
-          expect.objectContaining({
-            occupationLocations: ['GB-ENG', 'GB-SCT', 'GB-WLS', 'GB-NIR'],
-          }),
+          '/admin/professions/profession-id/versions/version-id/scope/edit',
         );
       });
     });
@@ -335,9 +105,6 @@ describe('TopLevelInformationController', () => {
       it('does not create a profession, and re-renders the top level information view with errors', async () => {
         const topLevelDetailsDtoWithNoAnswers = {
           name: '',
-          nations: undefined,
-          coversUK: null,
-          industries: undefined,
         };
 
         await controller.update(
@@ -354,18 +121,11 @@ describe('TopLevelInformationController', () => {
               name: {
                 text: 'professions.form.errors.name.empty',
               },
-              coversUK: {
-                text: 'professions.form.errors.nations.empty',
-              },
-              industries: {
-                text: 'professions.form.errors.industries.empty',
-              },
             },
           }),
         );
 
         expect(professionsService.save).not.toHaveBeenCalled();
-        expect(industriesService.all).toHaveBeenCalled();
       });
     });
 
@@ -376,26 +136,12 @@ describe('TopLevelInformationController', () => {
             id: 'profession-id',
           });
 
-          const version = professionVersionFactory.build({
-            profession: profession,
-          });
-
-          const industry = industryFactory.build({ id: 'construction-uuid' });
-
           professionsService.findWithVersions.mockResolvedValue(profession);
-          professionVersionsService.findWithProfession.mockResolvedValue(
-            version,
-          );
 
           const topLevelDetailsDtoWithChangeParam = {
             name: 'A new profession',
-            coversUK: '0',
-            nations: ['GB-ENG'],
-            industries: ['construction-uuid'],
             change: 'true',
           };
-
-          industriesService.findByIds.mockResolvedValue([industry]);
 
           await controller.update(
             topLevelDetailsDtoWithChangeParam,
@@ -416,25 +162,11 @@ describe('TopLevelInformationController', () => {
             id: 'profession-id',
           });
 
-          const version = professionVersionFactory.build({
-            profession: profession,
-          });
-
           professionsService.findWithVersions.mockResolvedValue(profession);
-          professionVersionsService.findWithProfession.mockResolvedValue(
-            version,
-          );
-
-          const industry = industryFactory.build({ id: 'construction-uuid' });
 
           const topLevelDetailsDtoWithoutChangeParam = {
             name: 'A new profession',
-            nations: ['GB-ENG'],
-            coversUK: '0',
-            industries: ['construction-uuid'],
           };
-
-          industriesService.findByIds.mockResolvedValue([industry]);
 
           await controller.update(
             topLevelDetailsDtoWithoutChangeParam,
@@ -444,7 +176,7 @@ describe('TopLevelInformationController', () => {
           );
 
           expect(response.redirect).toHaveBeenCalledWith(
-            '/admin/professions/profession-id/versions/version-id/regulatory-body/edit',
+            '/admin/professions/profession-id/versions/version-id/scope/edit',
           );
         });
       });
